Remove the `Value_Mut` type by unifying the register types
Not sure if this is the best move, but it doesn't seem like too much of
an issue at the moment.

+++ apps/ok/src/byte_code.rs
@@ -108,9 +108,9 @@ pub enum Micro_Op {
 
     Move(Register, Register),
 
-    Jump(usize),
-    Jump_Not_Zero(usize, Register),
+    Jump(u64),
+    Jump_Not_Zero(u64, Register),
 
-    Call(usize),
+    Call(u64),
     Return,
 }
